Refactor website using timlrx's next+Tailwind template
This commit is a complete refactor of the website source code using Timothy Lin's excellent next.js + Tailwind CSS starter template.

The code has been modified to work with the components and structure of the website. A license header was added for every file that was either used verbatim,
or had substantial similarities to the original source.

The structure was also refactored. Instead of semantic grouping of math / programming / blog, the content was split into a larger blog section, and photography.
Also, the tag system for timlrx's blog template was adapted to suit photography posts, however photo topics are not operational at this point.